Struct sn_routing::Node

pub struct Node { /* fields omitted */ }

Interface for sending and receiving messages to and from other nodes, in the role of a full sn_routing node.

A node is a part of the network that can route messages and be a member of a section or group location. Its methods can be used to send requests and responses as either an individual Node or as a part of a section or group location. Their src argument indicates that role, and can be any SrcLocation.

Implementations

impl Node

pub async fn new(config: NodeConfig) -> Result<(Self, EventStream)>

Create new node using the given config.

pub async fn public_key<'_>(&'_ self) -> PublicKey

Returns the PublicKey of this node.

pub async fn name<'_>(&'_ self) -> XorName

The name of this node.

pub async fn our_connection_info<'_>(&'_ self) -> Result<SocketAddr>

Returns connection info of this node.

pub async fn our_prefix<'_>(&'_ self) -> Option<Prefix>

Our Prefix once we are a part of the section.

pub async fn matches_our_prefix<'_, '_>(
&'_ self,
name: &'_ XorName
) -> Result<bool>

Finds out if the given XorName matches our prefix. Returns error if we don't have a prefix because we haven't joined any section yet.

pub async fn is_elder<'_>(&'_ self) -> bool

Returns whether the node is Elder.

pub async fn our_elders<'_>(&'_ self) -> Vec<Peer>

Returns the information of all the current section elders.

pub async fn our_elders_sorted_by_distance_to<'_, '_>(
&'_ self,
name: &'_ XorName
) -> Vec<Peer>

Returns the elders of our section sorted by their distance to name (closest first).

pub async fn our_adults<'_>(&'_ self) -> Vec<Peer>

Returns the information of all the current section adults.

pub async fn our_adults_sorted_by_distance_to<'_, '_>(
&'_ self,
name: &'_ XorName
) -> Vec<Peer>

Returns the adults of our section sorted by their distance to name (closest first). If we are not elder or if there are no adults in the section, returns empty vec.

pub async fn our_section<'_>(&'_ self) -> Option<EldersInfo>

Returns the info about our section or None if we are not joined yet.

pub async fn neighbour_sections<'_>(&'_ self) -> Vec<EldersInfo>

Returns the info about our neighbour sections.

pub async fn send_message<'_>(
&'_ mut self,
src: SrcLocation,
dst: DstLocation,
content: Bytes
) -> Result<()>

Send a message.

pub async fn send_message_to_client<'_>(
&'_ mut self,
peer_addr: SocketAddr,
msg: Bytes
) -> Result<()>

Send a message to a client peer.

pub async fn public_key_set<'_>(&'_ self) -> Result<PublicKeySet>

Returns the current BLS public key set or Error::InvalidState if we are not joined yet.

pub async fn secret_key_share<'_>(&'_ self) -> Result<SecretKeyShare>

Returns the current BLS secret key share or Error::InvalidState if we are not elder.

pub async fn our_history<'_>(&'_ self) -> Option<SectionProofChain>

Returns our section proof chain, or None if we are not joined yet.

pub async fn our_index<'_>(&'_ self) -> Result<usize>

Returns our index in the current BLS group or Error::InvalidState if section key was not generated yet.