Enum sn_routing::Event

pub enum Event {
    MessageReceived {
        content: Bytes,
        src: SrcLocation,
        dst: DstLocation,
        signed: Option<Signed>,
        section_pk: PublicKey,
    },
    MemberJoined {
        name: XorName,
        previous_name: Option<XorName>,
        age: u8,
    },
    MemberLeft {
        name: XorName,
        age: u8,
    },
    SectionSplit {
        elders: Elders,
        sibling_elders: Elders,
        self_status_change: NodeElderChange,
    },
    EldersChanged {
        elders: Elders,
        self_status_change: NodeElderChange,
    },
    RelocationStarted {
        previous_name: XorName,
    },
    Relocated {
        previous_name: XorName,
        new_keypair: Arc<Keypair>,
    },
    RestartRequired,
    ClientMsgReceived {
        msg: Box<ClientMsg>,
        user: EndUser,
    },
    ClientLost(SocketAddr),
    AdultsChanged {
        remaining: BTreeSet<XorName>,
        added: BTreeSet<XorName>,
        removed: BTreeSet<XorName>,
    },
}

An Event raised by a Node or Client via its event sender.

These are sent by sn_routing to the library’s user. It allows the user to handle requests and responses, and to react to changes in the network.

Request and Response events from section locations are only raised once the majority has been reached, i.e. enough members of the section have sent the same message.

Variants

MessageReceived

Received a message.

Fields of MessageReceived

content: Bytes

The content of the message.

src: SrcLocation

The source location that sent the message.

dst: DstLocation

The destination location that receives the message.

signed: Option<Signed>

The signed if the message was set to be aggregated at source.

section_pk: PublicKey

The Sender’s Section PK.

MemberJoined

A new peer joined our section.

Fields of MemberJoined

name: XorName

Name of the node

previous_name: Option<XorName>

Previous name before relocation or None if it is a new node.

age: u8

Age of the node

MemberLeft

A node left our section.

Fields of MemberLeft

name: XorName

Name of the node

age: u8

Age of the node

SectionSplit

Our section has split.

Fields of SectionSplit

elders: Elders

The Elders of our section.

sibling_elders: Elders

The Elders of the sibling section.

self_status_change: NodeElderChange

Promoted, demoted or no change?

EldersChanged

The set of elders in our section has changed.

Fields of EldersChanged

elders: Elders

The Elders of our section.

self_status_change: NodeElderChange

Promoted, demoted or no change?

RelocationStarted

This node has started relocating to other section. Will be followed by Relocated when the node finishes joining the destination section.

Fields of RelocationStarted

previous_name: XorName

Previous name before relocation

Relocated

This node has completed relocation to other section.

Fields of Relocated

previous_name: XorName

Old name before the relocation.

new_keypair: Arc<Keypair>

New keypair to be used after relocation.

RestartRequired

Disconnected or failed to connect - restart required.

ClientMsgReceived

Received a message from a client node.

Fields of ClientMsgReceived

msg: Box<ClientMsg>

The content of the message.

user: EndUser

The SocketAddr and PublicKey that sent the message. (Note: socket_id will be a random hash, to map against the actual socketaddr)

ClientLost(SocketAddr)

Failed in sending a message to client, or connection to client is lost

AdultsChanged

Notify the current list of adult nodes, in case of churning.

Fields of AdultsChanged

remaining: BTreeSet<XorName>

Remaining Adults in our section.

added: BTreeSet<XorName>

New Adults in our section.

removed: BTreeSet<XorName>

Removed Adults in our section.