Struct sn_routing::Routing

pub struct Routing { /* fields omitted */ }

Interface for sending and receiving messages to and from other nodes, in the role of a full routing node.

A node is a part of the network that can route messages and be a member of a section or group location. Its methods can be used to send requests and responses as either an individual Node or as a part of a section or group location. Their src argument indicates that role, and can be any SrcLocation.

Implementations

impl Routing

pub async fn new(config: Config) -> Result<(Self, EventStream)>

Creates new node using the given config and bootstraps it to the network.

NOTE: It's not guaranteed this function ever returns. This can happen due to messages being lost in transit during bootstrapping, or other reasons. It's the responsibility of the caller to handle this case, for example by using a timeout.

pub async fn set_joins_allowed<'_>(&'_ self, joins_allowed: bool) -> Result<()>

Sets the JoinsAllowed flag.

pub async fn age<'_>(&'_ self) -> u8

Returns the current age of this node.

pub async fn public_key<'_>(&'_ self) -> PublicKey

Returns the ed25519 public key of this node.

pub async fn sign<'_, '_>(&'_ self, data: &'_ [u8]) -> Signature

Signs any data with the ed25519 key of this node.

pub async fn verify<'_, '_, '_>(
&'_ self,
data: &'_ [u8],
signature: &'_ Signature
) -> bool

Verifies signature on data with the ed25519 public key of this node.

pub async fn name<'_>(&'_ self) -> XorName

The name of this node.

pub async fn our_connection_info<'_>(&'_ self) -> Result<SocketAddr>

Returns connection info of this node.

pub async fn our_prefix<'_>(&'_ self) -> Prefix

Prefix of our section

pub async fn matches_our_prefix<'_, '_>(&'_ self, name: &'_ XorName) -> bool

Finds out if the given XorName matches our prefix.

pub async fn is_elder<'_>(&'_ self) -> bool

Returns whether the node is Elder.

pub async fn our_elders<'_>(&'_ self) -> Vec<Peer>

Returns the information of all the current section elders.

pub async fn our_elders_sorted_by_distance_to<'_, '_>(
&'_ self,
name: &'_ XorName
) -> Vec<Peer>

Returns the elders of our section sorted by their distance to name (closest first).

pub async fn our_adults<'_>(&'_ self) -> Vec<Peer>

Returns the information of all the current section adults.

pub async fn our_adults_sorted_by_distance_to<'_, '_>(
&'_ self,
name: &'_ XorName
) -> Vec<Peer>

Returns the adults of our section sorted by their distance to name (closest first). If we are not elder or if there are no adults in the section, returns empty vec.

pub async fn our_section<'_>(&'_ self) -> EldersInfo

Returns the info about our section or None if we are not joined yet.

pub async fn neighbour_sections<'_>(&'_ self) -> Vec<EldersInfo>

Returns the info about our neighbour sections.

pub async fn match_section<'_, '_>(
&'_ self,
name: &'_ XorName
) -> (Option<PublicKey>, Option<EldersInfo>)

Returns the info about the section matches the name.

pub async fn send_message<'_>(
&'_ self,
src: SrcLocation,
dst: DstLocation,
content: Bytes
) -> Result<()>

Send a message. Messages sent here, either section to section or node to node are signed and validated upon receipt by routing itself.

pub async fn send_message_to_client<'_>(
&'_ self,
recipient: SocketAddr,
message: Bytes
) -> Result<()>

Send a message to a client peer. Messages sent to a client are not signed or validated as part of the routing library.

pub async fn public_key_set<'_>(&'_ self) -> Result<PublicKeySet>

Returns the current BLS public key set if this node has one, or Error::InvalidState otherwise.

pub async fn secret_key_share<'_>(&'_ self) -> Result<SecretKeyShare>

Returns the current BLS secret key share or Error::InvalidState if we are not elder.

pub async fn sign_with_secret_key_share<'_, '_>(
&'_ self,
data: &'_ [u8]
) -> Result<SignatureShare>

Signs data with the BLS secret key share of this node, if it has any. Returns Error::MissingSecretKeyShare otherwise.

pub async fn our_history<'_>(&'_ self) -> SectionProofChain

Returns our section proof chain.

pub async fn our_index<'_>(&'_ self) -> Result<usize>

Returns our index in the current BLS group if this node is a member of one, or Error::MissingSecretKeyShare otherwise.