Struct sn_routing::Routing

pub struct Routing { /* fields omitted */ }

Interface for sending and receiving messages to and from other nodes, in the role of a full routing node.

A node is a part of the network that can route messages and be a member of a section or group location. Its methods can be used to send requests and responses as either an individual Node or as a part of a section or group location. Their src argument indicates that role, and can be sn_messaging::SrcLocation::Node or sn_messaging::SrcLocation::Section.

Implementations

impl Routing

pub async fn new(config: Config) -> Result<(Self, EventStream)>

Creates new node using the given config and bootstraps it to the network.

NOTE: It’s not guaranteed this function ever returns. This can happen due to messages being lost in transit during bootstrapping, or other reasons. It’s the responsibility of the caller to handle this case, for example by using a timeout.

pub async fn set_joins_allowed(&self, joins_allowed: bool) -> Result<()>

Sets the JoinsAllowed flag.

pub async fn propose_offline(&self, name: XorName) -> Result<()>

Starts a proposal that a node has gone offline. This can be done only by an Elder.

pub async fn age(&self) -> u8

Returns the current age of this node.

pub async fn public_key(&self) -> PublicKey

Returns the ed25519 public key of this node.

pub async fn keypair_as_bytes(&self) -> [u8; 64]

Returns the ed25519 keypair of this node, as bytes.

pub async fn sign_as_node(&self, data: &[u8]) -> Signature

Signs data with the ed25519 key of this node.

pub async fn sign_as_elder(
&self,
data: &[u8],
public_key: &PublicKey
) -> Result<SignatureShare>

Signs data with the BLS secret key share of this node, if it has any. Returns Error::MissingSecretKeyShare otherwise.

pub async fn verify(&self, data: &[u8], signature: &Signature) -> bool

Verifies signature on data with the ed25519 public key of this node.

pub async fn name(&self) -> XorName

The name of this node.

pub fn our_connection_info(&self) -> SocketAddr

Returns connection info of this node.

pub async fn section_chain(&self) -> SectionChain

Returns the Section Proof Chain

pub async fn our_prefix(&self) -> Prefix

Prefix of our section

pub async fn matches_our_prefix(&self, name: &XorName) -> bool

Finds out if the given XorName matches our prefix.

pub async fn is_elder(&self) -> bool

Returns whether the node is Elder.

pub async fn our_elders(&self) -> Vec<Peer>

Returns the information of all the current section elders.

pub async fn our_elders_sorted_by_distance_to(
&self,
name: &XorName
) -> Vec<Peer>

Returns the elders of our section sorted by their distance to name (closest first).

pub async fn our_adults(&self) -> Vec<Peer>

Returns the information of all the current section adults.

pub async fn our_adults_sorted_by_distance_to(
&self,
name: &XorName
) -> Vec<Peer>

Returns the adults of our section sorted by their distance to name (closest first). If we are not elder or if there are no adults in the section, returns empty vec.

pub async fn our_section(&self) -> SectionAuthorityProvider

Returns the info about our section or None if we are not joined yet.

pub async fn other_sections(&self) -> Vec<SectionAuthorityProvider>

Returns the info about other sections in the network known to us.

pub async fn section_key(&self, prefix: &Prefix) -> Option<PublicKey>

Returns the last known public key of the section with prefix.

pub async fn matching_section(
&self,
name: &XorName
) -> (Option<PublicKey>, Option<SectionAuthorityProvider>)

Returns the info about the section matching the name.

pub async fn send_message(
&self,
itinerary: Itinerary,
content: Bytes,
additional_proof_chain_key: Option<PublicKey>
) -> Result<()>

Send a message. Messages sent here, either section to section or node to node are signed and validated upon receipt by routing itself.

additional_proof_chain_key is a key to be included in the proof chain attached to the message. This is useful when the message contains some data that is signed with a different key than the whole message is so that the recipient can verify such key.

pub async fn public_key_set(&self) -> Result<PublicKeySet>

Returns the current BLS public key set if this node has one, or Error::InvalidState otherwise.

pub async fn our_history(&self) -> SectionChain

Returns our section proof chain.

pub async fn our_index(&self) -> Result<usize>

Returns our index in the current BLS group if this node is a member of one, or Error::MissingSecretKeyShare otherwise.